Nearly operational derive

+++ uclicious-example/src/main.rs
@@ -1,9 +1,13 @@
 use uclicious_derive::Uclicious;
+use uclicious;
 
 #[derive(Debug,Uclicious)]
 pub struct Connection {
     host: String,
-    port: u32,
+    #[ucl(default = "420")]
+    port: i64,
+//    #[ucl(path = "type")]
+//    kind: String,
 }
 fn main() {
     println!("Hello, world!");
